refactor: useAuthReady TS로 전환 및 중복 디코딩 로직 제거
